# open-api/ram-openapi-specs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/admin-sign-up": {
      "post": {
        "operationId": "signUpAdmin",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/change-password": {
      "post": {
        "operationId": "changePassword",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCredentialDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/validateToken": {
      "post": {
        "operationId": "validateAccessToken",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateAccessTokenResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/shared-link": {
      "get": {
        "operationId": "getAllSharedLinks",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SharedLinkResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Shared Link"
        ]
      },
      "post": {
        "operationId": "createSharedLink",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SharedLinkCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SharedLinkResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Shared Link"
        ]
      }
    },
    "/shared-link/me": {
      "get": {
        "operationId": "getMySharedLink",
        "parameters": [
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "required": false,
            "in": "query",
            "example": "password",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SharedLinkResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Shared Link"
        ]
      }
    },
    "/shared-link/{id}": {
      "delete": {
        "operationId": "removeSharedLink",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Shared Link"
        ]
      },
      "get": {
        "operationId": "getSharedLinkById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SharedLinkResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Shared Link"
        ]
      },
      "patch": {
        "operationId": "updateSharedLink",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SharedLinkEditDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SharedLinkResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Shared Link"
        ]
      }
    },
    "/shared-link/{id}/assets": {
      "delete": {
        "operationId": "removeSharedLinkAssets",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetIdsDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AssetIdsResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Shared Link"
        ]
      },
      "put": {
        "operationId": "addSharedLinkAssets",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetIdsDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AssetIdsResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Shared Link"
        ]
      }
    },
...
  },
  "info": {
    "title": "ram",
    "description": "ram API",
    "version": "1.103.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "Bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ram_access_token"
      },
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "AlbumResponseDto": {
        "properties": {
          "albumName": {
            "type": "string"
          },
          "albumThumbnailAssetId": {
            "nullable": true,
            "type": "string"
          },
          "albumUsers": {
            "items": {
              "$ref": "#/components/schemas/AlbumUserResponseDto"
            },
            "type": "array"
          },
          "assetCount": {
            "type": "integer"
          },
          "assets": {
            "items": {
              "$ref": "#/components/schemas/AssetResponseDto"
            },
            "type": "array"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "endDate": {
            "format": "date-time",
            "type": "string"
          },
          "hasSharedLink": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "isActivityEnabled": {
            "type": "boolean"
          },
          "lastModifiedAssetTimestamp": {
            "format": "date-time",
            "type": "string"
          },
          "order": {
            "$ref": "#/components/schemas/AssetOrder"
          },
          "owner": {
            "$ref": "#/components/schemas/UserResponseDto"
          },
          "ownerId": {
            "type": "string"
          },
          "shared": {
            "type": "boolean"
          },
          "sharedUsers": {
            "deprecated": true,
            "description": "This property was deprecated in v1.102.0",
            "items": {
              "$ref": "#/components/schemas/UserResponseDto"
            },
            "type": "array"
          },
          "startDate": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "albumName",
          "albumThumbnailAssetId",
          "albumUsers",
          "assetCount",
          "assets",
          "createdAt",
          "description",
          "hasSharedLink",
          "id",
          "isActivityEnabled",
          "owner",
          "ownerId",
          "shared",
          "sharedUsers",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "AlbumUserResponseDto": {
        "properties": {
          "role": {
            "$ref": "#/components/schemas/AlbumUserRole"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponseDto"
          }
        },
        "required": [
          "role",
          "user"
        ],
        "type": "object"
      },
      "AlbumUserRole": {
        "enum": [
          "editor",
          "viewer"
        ],
        "type": "string"
      },
...
      "AssetIdsDto": {
        "properties": {
          "assetIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "assetIds"
        ],
        "type": "object"
      },
      "AssetIdsResponseDto": {
        "properties": {
          "assetId": {
            "type": "string"
          },
          "error": {
            "enum": [
              "duplicate",
              "no_permission",
              "not_found"
            ],
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "assetId",
          "success"
        ],
        "type": "object"
      },
...
      "AssetOrder": {
        "enum": [
          "asc",
          "desc"
        ],
        "type": "string"
      },
      "AssetResponseDto": {
        "properties": {
          "checksum": {
            "description": "base64 encoded sha1 hash",
            "type": "string"
          },
          "deviceAssetId": {
            "type": "string"
          },
          "deviceId": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "exifInfo": {
            "$ref": "#/components/schemas/ExifResponseDto"
          },
          "fileCreatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "fileModifiedAt": {
            "format": "date-time",
            "type": "string"
          },
          "hasMetadata": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "isArchived": {
            "type": "boolean"
          },
          "isExternal": {
            "deprecated": true,
            "description": "This property was deprecated in v1.104.0",
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "isOffline": {
            "type": "boolean"
          },
          "isReadOnly": {
            "deprecated": true,
            "description": "This property was deprecated in v1.104.0",
            "type": "boolean"
          },
          "isTrashed": {
            "type": "boolean"
          },
          "libraryId": {
            "type": "string"
          },
          "livePhotoVideoId": {
            "nullable": true,
            "type": "string"
          },
          "localDateTime": {
            "format": "date-time",
            "type": "string"
          },
          "originalFileName": {
            "type": "string"
          },
          "originalPath": {
            "type": "string"
          },
          "owner": {
            "$ref": "#/components/schemas/UserResponseDto"
          },
          "ownerId": {
            "type": "string"
          },
          "people": {
            "items": {
              "$ref": "#/components/schemas/PersonWithFacesResponseDto"
            },
            "type": "array"
          },
          "resized": {
            "type": "boolean"
          },
          "smartInfo": {
            "$ref": "#/components/schemas/SmartInfoResponseDto"
          },
          "stack": {
            "items": {
              "$ref": "#/components/schemas/AssetResponseDto"
            },
            "type": "array"
          },
          "stackCount": {
            "nullable": true,
            "type": "integer"
          },
          "stackParentId": {
            "nullable": true,
            "type": "string"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagResponseDto"
            },
            "type": "array"
          },
          "thumbhash": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/AssetTypeEnum"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "checksum",
          "deviceAssetId",
          "deviceId",
          "duration",
          "fileCreatedAt",
          "fileModifiedAt",
          "hasMetadata",
          "id",
          "isArchived",
          "isFavorite",
          "isOffline",
          "isTrashed",
          "libraryId",
          "localDateTime",
          "originalFileName",
          "originalPath",
          "ownerId",
          "resized",
          "stackCount",
          "thumbhash",
          "type",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "AssetTypeEnum": {
        "enum": [
          "IMAGE",
          "VIDEO",
          "AUDIO",
          "OTHER"
        ],
        "type": "string"
      },
...
      "ChangePasswordDto": {
        "properties": {
          "newPassword": {
            "example": "password",
            "minLength": 8,
            "type": "string"
          },
          "password": {
            "example": "password",
            "type": "string"
          }
        },
        "required": [
          "newPassword",
          "password"
        ],
        "type": "object"
      },
...
      "LoginCredentialDto": {
        "properties": {
          "email": {
            "example": "testuser@email.com",
            "type": "string"
          },
          "password": {
            "example": "password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
      "LoginResponseDto": {
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "isAdmin": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "profileImagePath": {
            "type": "string"
          },
          "shouldChangePassword": {
            "type": "boolean"
          },
          "userEmail": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "isAdmin",
          "name",
          "profileImagePath",
          "shouldChangePassword",
          "userEmail",
          "userId"
        ],
        "type": "object"
      },
      "LogoutResponseDto": {
        "properties": {
          "redirectUri": {
            "type": "string"
          },
          "successful": {
            "type": "boolean"
          }
        },
        "required": [
          "redirectUri",
          "successful"
        ],
        "type": "object"
      },
...
      "PersonWithFacesResponseDto": {
        "properties": {
          "birthDate": {
            "format": "date",
            "nullable": true,
            "type": "string"
          },
          "faces": {
            "items": {
              "$ref": "#/components/schemas/AssetFaceWithoutPersonResponseDto"
            },
            "type": "array"
          },
          "id": {
            "type": "string"
          },
          "isHidden": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "thumbnailPath": {
            "type": "string"
          }
        },
        "required": [
          "birthDate",
          "faces",
          "id",
          "isHidden",
          "name",
          "thumbnailPath"
        ],
        "type": "object"
      },
...
      "SharedLinkCreateDto": {
        "properties": {
          "albumId": {
            "format": "uuid",
            "type": "string"
          },
          "allowDownload": {
            "default": true,
            "type": "boolean"
          },
          "allowUpload": {
            "type": "boolean"
          },
          "assetIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          },
          "description": {
            "type": "string"
          },
          "expiresAt": {
            "default": null,
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "showMetadata": {
            "default": true,
            "type": "boolean"
          },
          "type": {
            "$ref": "#/components/schemas/SharedLinkType"
          }
        },
        "required": [
          "type"
        ],
        "type": "object"
      },
      "SharedLinkEditDto": {
        "properties": {
          "allowDownload": {
            "type": "boolean"
          },
          "allowUpload": {
            "type": "boolean"
          },
          "changeExpiryTime": {
            "description": "Few clients cannot send null to set the expiryTime to never.\nSetting this flag and not sending expiryAt is considered as null instead.\nClients that can send null values can ignore this.",
            "type": "boolean"
          },
          "description": {
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "showMetadata": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "SharedLinkResponseDto": {
        "properties": {
          "album": {
            "$ref": "#/components/schemas/AlbumResponseDto"
          },
          "allowDownload": {
            "type": "boolean"
          },
          "allowUpload": {
            "type": "boolean"
          },
          "assets": {
            "items": {
              "$ref": "#/components/schemas/AssetResponseDto"
            },
            "type": "array"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "nullable": true,
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "password": {
            "nullable": true,
            "type": "string"
          },
          "showMetadata": {
            "type": "boolean"
          },
          "token": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/SharedLinkType"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "allowDownload",
          "allowUpload",
          "assets",
          "createdAt",
          "description",
          "expiresAt",
          "id",
          "key",
          "password",
          "showMetadata",
          "type",
          "userId"
        ],
        "type": "object"
      },
      "SharedLinkType": {
        "enum": [
          "ALBUM",
          "INDIVIDUAL"
        ],
        "type": "string"
      },
      "SignUpDto": {
        "properties": {
          "email": {
            "example": "testuser@email.com",
            "type": "string"
          },
          "name": {
            "example": "Admin",
            "type": "string"
          },
          "password": {
            "example": "password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "name",
          "password"
        ],
        "type": "object"
      },
      "SmartInfoResponseDto": {
        "properties": {
          "objects": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "UserAvatarColor": {
        "enum": [
          "primary",
          "pink",
          "red",
          "yellow",
          "blue",
          "green",
          "purple",
          "orange",
          "gray",
          "amber"
        ],
        "type": "string"
      },
...
      "UserResponseDto": {
        "properties": {
          "avatarColor": {
            "$ref": "#/components/schemas/UserAvatarColor"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "isAdmin": {
            "type": "boolean"
          },
          "memoriesEnabled": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "oauthId": {
            "type": "string"
          },
          "profileImagePath": {
            "type": "string"
          },
          "quotaSizeInBytes": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "quotaUsageInBytes": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "shouldChangePassword": {
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "storageLabel": {
            "nullable": true,
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "avatarColor",
          "createdAt",
          "deletedAt",
          "email",
          "id",
          "isAdmin",
          "name",
          "oauthId",
          "profileImagePath",
          "quotaSizeInBytes",
          "quotaUsageInBytes",
          "shouldChangePassword",
          "status",
          "storageLabel",
          "updatedAt"
        ],
        "type": "object"
      },
      "UserStatus": {
        "enum": [
          "active",
          "removing",
          "deleted"
        ],
        "type": "string"
      },
      "ValidateAccessTokenResponseDto": {
        "properties": {
          "authStatus": {
            "type": "boolean"
          }
        },
        "required": [
          "authStatus"
        ],
        "type": "object"
      },
...
    }
  }
}